==> notebooks/1.0-ak-extract-titanic-data.ipynb <==
# Extracting Titanic Disaster Data From Kaggle

!pip install python-dotenv

from dotenv import load_dotenv, find_dotenv

# search for .env until finding it
dotenv_path = find_dotenv()
# load up entries as environment variables
load_dotenv(dotenv_path)

import os
KAGGLE_USERNAME = os.environ.get("KAGGLE_USERNAME")
print(KAGGLE_USERNAME)

import requests
from requests import session
import os
from dotenv import load_dotenv, find_dotenv

payload = {
    'action': 'login',
    'username': os.environ.get("KAGGLE_USERNAME"),
    'password': os.environ.get("KAGGLE_PASSWORD")
}
url = 'https://storage.googleapis.com/kagglesdsdata/competitions/3136/26502/train.csv?GoogleAccessId=web-data@kaggle-161607.iam.gserviceaccount.com&Expires=1603494299&Signature=hQE3WV7tE6fghttLoy6T8FaxguHicpYZTdgPEpkzB2qJx6EFA%2F2QmRORbhbrkIj8Op3DFxzuHZVRbB%2Fu%2Bbb%2BU7iSW0jPmw7dBiH3VO%2BcAqxLF1eXOFs7dPhEVWHP0k4qT4x0bmDLpLqROKSikti3I%2Bx8xSgyThDhTGm50%2BOB9MnldxnrqbQexZNL6OZrDomkM59Zdz8HNopjWd5%2FNOBANEsbGY1VM5ELSWrGHXUOK3vf21zcp5houx0eB%2Fm6u1j0SynuQS8aovuzA4BJSpvj5ZQ06ydTJJKFglW9N4cjfr0yhMJ4YlO9F0QnKLUftNhvSWzMkNXhanwVw%2FOVfxtDrg%3D%3D&response-content-disposition=attachment%3B+filename%3Dtrain.csv'

# setup session
with session() as c:
    c.post('https://www.kaggle.com/account/login', data=payload)
    response = c.get(url)
    print(response.text)

from requests import session
payload = {
    'action': 'login',
    'username': os.environ.get("KAGGLE_USERNAME"),
    'password': os.environ.get("KAGGLE_PASSWORD")
}

def extract_data(url, file_path):
    with session() as c:
        c.post('https://www.kaggle.com/account/login', data=payload)
        with open(file_path, 'wb') as handle: # wb so that it is binary
            response = c.get(url, stream=True)
            for block in response.iter_content(1024):
                handle.write(block)


train_url = 'https://storage.googleapis.com/kagglesdsdata/competitions/3136/26502/train.csv?GoogleAccessId=web-data@kaggle-161607.iam.gserviceaccount.com&Expires=1603494299&Signature=hQE3WV7tE6fghttLoy6T8FaxguHicpYZTdgPEpkzB2qJx6EFA%2F2QmRORbhbrkIj8Op3DFxzuHZVRbB%2Fu%2Bbb%2BU7iSW0jPmw7dBiH3VO%2BcAqxLF1eXOFs7dPhEVWHP0k4qT4x0bmDLpLqROKSikti3I%2Bx8xSgyThDhTGm50%2BOB9MnldxnrqbQexZNL6OZrDomkM59Zdz8HNopjWd5%2FNOBANEsbGY1VM5ELSWrGHXUOK3vf21zcp5houx0eB%2Fm6u1j0SynuQS8aovuzA4BJSpvj5ZQ06ydTJJKFglW9N4cjfr0yhMJ4YlO9F0QnKLUftNhvSWzMkNXhanwVw%2FOVfxtDrg%3D%3D&response-content-disposition=attachment%3B+filename%3Dtrain.csv'
test_url = 'https://storage.googleapis.com/kagglesdsdata/competitions/3136/26502/test.csv?GoogleAccessId=web-data@kaggle-161607.iam.gserviceaccount.com&Expires=1603494701&Signature=hdTFEQtEIzWFtPSDNH5e%2FXf1uEcMehvsz8X6EUvWd84U6wVXrC46xuT4sg2kukX85JwQCDJDahEi0P2Vi8oU6CYFz6ZWjJji42xSVi1fbUUBSRSeyQ0yflM2ctZpUmqU%2BsRBkBRspsS1dO1v3IqOOpowhzmn%2FtIF8Behkujzywgk6tINxoo1Rk1uYuS0O9VGYC167anXump3wa7H26er00hGCifsehGyk8c8y4shl0ZXxWEQD17UXRn5MmpVBS3jq6AG3xmEdlgt%2BWgn77%2B3Y3VXN0dhjgdZxU%2FHQXPBFjHKJgdCG7gI%2B2dolZZWaK5Dl4b0k36NMTv3sCNPurDHsw%3D%3D&response-content-disposition=attachment%3B+filename%3Dtest.csv'

raw_data_path = os.path.join(os.path.pardir, 'data', 'raw')
train_data_path = os.path.join(raw_data_path, 'train.csv')
test_data_path = os.path.join(raw_data_path, 'test.csv')

extract_data(train_url, train_data_path)
extract_data(test_url, test_data_path)

## Building the file script

get_raw_data_script_file = os.path.join(os.path.pardir, 'src', 'data', 'get_raw_data.py')

%%writefile $get_raw_data_script_file

import os
from dotenv import find_dotenv, load_dotenv
from requests import session
import logging

payload = {
    'action': 'login',
    'username': os.environ.get("KAGGLE_USERNAME"),
    'password': os.environ.get("KAGGLE_PASSWORD")
}

def extract_data(url, file_path):
    with session() as c:
        c.post('https://www.kaggle.com/account/login', data=payload)
        with open(file_path, 'wb') as handle: # wb so that it is binary
            response = c.get(url, stream=True)
            for block in response.iter_content(1024):
                handle.write(block)

def main(project_dir):
    
    logger = logging.getLogger(__name__)
    logger.info('getting raw data')
    
    train_url = os.environ.get("TRAIN_URL")
    test_url = os.environ.get("TEST_URL")
                
    raw_data_path = os.path.join(os.path.pardir, 'data', 'raw')
    train_data_path = os.path.join(raw_data_path, 'train.csv')
    test_data_path = os.path.join(raw_data_path, 'test.csv')

    extract_data(train_url, train_data_path)
    extract_data(test_url, test_data_path)
    logger.info('downloaded raw data training and test data')

if __name__ == '__main__':
    project_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
    log_fmt = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    logging.basicConfig(level=logging.INFO, format=log_fmt)
    
    dotenv_path = find_dotenv()
    load_dotenv(dotenv_path)
    
    # call the main
    main(project_dir)

!python $get_raw_data_script_file
